requirements.txt
asgiref==3.2.10
cairocffi==1.2.0
CairoSVG==2.5.0
cffi==1.14.3
colorama==0.4.4
cssselect2==0.4.1
Django==3.1.4
django-admin-sortable2==0.7.7
django-appconf==1.0.4
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cleanup==5.1.0
django-form-admin==0.5.1
django-front==0.5.14
django-image-cropping==1.5.0
django-js-asset==1.2.2
django-modeladmin-reorder==0.3.1
django-modeltranslation==0.16
django-unused-media==0.2.2
easy-thumbnails==2.7
html5lib==1.1
pdfkit==0.6.1
Pillow==8.0.1
pycparser==2.20
Pyphen==0.10.0
pytz==2020.1
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
WeasyPrint==52
webencodings==0.5.1
xlrd==1.2.0
